+++ pyrlang/dist_proto/server.py
@@ -21,8 +21,8 @@ import logging
 import random
 import struct
 
-from pyrlang2.dist_proto import version
-from pyrlang2.dist_proto.base_dist_protocol import BaseDistProtocol, \
+from pyrlang.dist_proto import version
+from pyrlang.dist_proto.base_dist_protocol import BaseDistProtocol, \
     DistributionError
 from term import util
